requirements.txt
boto==2.48.0
boto3==1.4.7
botocore==1.7.43
Django==1.11
django-storages==1.6.5
djangorestframework==3.7.3
docutils==0.14
jmespath==0.9.3
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.11
six==1.11.0
